requirements.txt
/tmp/airflow_provider_weaviate-0.0.1-py3-none-any.whl
apache-airflow-providers-github==2.3.1
apache-airflow-providers-slack==7.3.2
weaviate-client==3.22.0
streamlit==1.25.0
html2text==2020.1.16
pypandoc==1.11
langchain==0.0.240
markdownify==0.11.6
stackapi==0.2.0
protobuf==3.20.1
openai==0.28.1
transformers==4.28.0
tokenizers==0.13.3
torch==2.0.0
pyarrow==13.0.0
goose3==3.1.17